+++ src/Serilog.Sinks.RollingFileAlternate/Sinks/HourlyRolling/HourlyLogFileInfo.cs
@@ -2,40 +2,40 @@
 using System.IO;
 using System.Text.RegularExpressions;
 
-namespace Serilog.Sinks.RollingFileAlternate.Sinks.SizeRollingFileSink
+namespace Serilog.Sinks.RollingFileAlternate.Sinks.HourlyRolling
 {
-    internal class LogFileInfo
+    internal class HourlyLogFileInfo
     {
         private const string NumberFormat = "00000";
-        private const string DateFormat = "yyyyMMdd";
+        private const string DateFormat = "HH";
 
         internal uint Sequence { get; private set; }
         internal string FileName { get; private set; }
         internal DateTime Date { get; private set; }
 
-        public LogFileInfo(DateTime date, uint sequence)
+        public HourlyLogFileInfo(DateTime date, uint sequence)
         {
             this.Date = date;
             this.Sequence = sequence;
             this.FileName = String.Format("{0}-{1}.log", date.ToString(DateFormat), sequence.ToString(NumberFormat));
         }
 
-        public LogFileInfo Next()
+        public HourlyLogFileInfo Next()
         {
             DateTime now = DateTime.UtcNow;
-            if (this.Date.Date != now.Date)
+            if (this.Date.Hour != now.Hour)
             {
-                return new LogFileInfo(now, 1);
+                return new HourlyLogFileInfo(now, 1);
             }
 
-            return new LogFileInfo(now, this.Sequence + 1);
+            return new HourlyLogFileInfo(now, this.Sequence + 1);
         }
 
-        internal static LogFileInfo GetLatestOrNew(DateTime date, string logDirectory)
+        internal static HourlyLogFileInfo GetLatestOrNew(DateTime date, string logDirectory)
         {
             string pattern = date.ToString(DateFormat) + @"-(\d{5}).log";
 
-            var logFileInfo = new LogFileInfo(date, 1);
+            var logFileInfo = new HourlyLogFileInfo(date, 1);
 
             foreach (var filePath in Directory.GetFiles(logDirectory))
             {
@@ -46,7 +46,7 @@ namespace Serilog.Sinks.RollingFileAlternate.Sinks.SizeRollingFileSink
 
                     if (sequence > logFileInfo.Sequence)
                     {
-                        logFileInfo = new LogFileInfo(date, sequence);
+                        logFileInfo = new HourlyLogFileInfo(date, sequence);
                     }
                 }
             }