+++ test/Serilog.Sinks.RollingFileAlternate.Tests/SizeLimitReachedTests.cs
@@ -14,7 +14,7 @@ namespace Serilog.Sinks.RollingFileAlternate.Tests
         public void ReachedWhenAmountOfCharactersWritten()
         {
             var formatter = new RawFormatter();
-            var components = new LogFileInfo(new DateTime(2015, 01, 15), 0);
+            var components = new SizeLimitedLogFileInfo(new DateTime(2015, 01, 15), 0);
             var logFile = new SizeLimitedLogFileDescription(components, 1);
             using (var str = new MemoryStream())
             using (var wr = new StreamWriter(str, Encoding.UTF8))